Lesson 244
Casting Lists
Cast the list retrieved from fetchTopIds to int so that dart knows that
it is a List of integers.

+++ Flutter_and_Dart/news/lib/src/resources/news_api_provider.dart
@@ -13,7 +13,7 @@ class NewsApiProvider {
         .get(Uri.parse('$_root/topstories.json'));
     final ids = jsonDecode(response.body);
 
-    return ids;
+    return ids.cast<int>();
   }
 
   Future<ItemModel> fetchItem(int id) async {
